dbbackups: Update grants for misc hosts other than m1
Update grants for more granularity ones after mariadb 10.6 upgrade,
as well as the right set of hosts (dbprov) that produces those
logical backups.

m1 section was done on a previous patch.

Sections affected:

* m2
* m3
* m5
* db_inventory
* analytics_meta
* backup1-eqiad
* backup1-codfw

* es hosts (only the grant granularity, as those are taken
  from backups hosts, not dbprov hosts)

The only other thing missing will be the m1 production grants
for statistics gathering (out of scope, not the user for generating
backups).

Bug: T383902
Change-Id: I61a09d7b4fe9a6b987bfe3c97520303b9c7ed807