trafficserver: use mobileapps directly for hewiki APIs
In order to test the viability of directly routing to mobileapps/pcs
without restbase, using mobileapps's self-managed cassandra cache,
route hewiki's mobile-html API path to the rest-gateway which in turn
routes to the mobileapps/PCS k8s service.

changeprop support is already in place for pregeneration without
restbase for hewiki.

Bug: T372746
Change-Id: I200a7a6b193adee03bee9c6991d4a1a5d040ae0f