conftool-data: add wikikube workers to kartotherian-k8s-ssl
We have deployed kartotherian on Wikikube, and compared latency
and tile quality with the bare metal hosts. We are now ready
to add Wikikube workers behind the kartotherian-k8s-ssl LVS
endpoint, to slowly add traffic to the k8s pods.

Assumption: any wikikube worker is fine since kube-proxy
will route the traffic to the hosts running the right pods/containers.

Once the migration is over we'll likely move Kartotherian
to Istio ingress.

Bug: T216826
Change-Id: Ia5ac70905c5d736a30672de137eec433c6239803