jobrunner: remove PHP major version httpbb assertion
This was added in Ia72fcea28ab2e4bcb717e758a0820f68a6a0eee1, as part of
the 7.2 -> 7.4 migration.

Later, in Ie922aa19454532801619f25705022deba41f87b9, the bare-metal-only
force_php assertions were removed, but the assertion on X-Powered-By was
not.

One might argue that the latter is useful, in that it guarantees the
request made it down into PHP, but this is less useful now that a more
distinctive response code (422) is returned (was: 500) and implicitly
validates that it was indeed RunSingleJob.php that handled the request.

In any case, I think it makes sense to remove this header assertion
rather than relax it (e.g., to ^PHP/).

Bug: T383845
Change-Id: Ic7dde595718f67adc915869e6c416e466dfbf781