varnish: create new policy that allows websockets but also caches
As explained by taavi in T274228#10529569:

The current caching policy is that if "websockets" is used then
upgrading to websockets is supported and other requests are NOT cached
(pass).

This policy is shared between Phorge (Phabricator) and Etherpad.

But if we want Phorge to cache requests and also still allow
websockets (for aphlict / realtime notifications) then we need to
create a new policy here that combines both.

Then it could be used by Phorge without affecting Etherpad.

Bug: T274228
Change-Id: Ic4e5da4e2a830dc92052a208c796a209f5b6d272