wcqs: proxy requests to query qui to new wikikube endpoint
The new query gui is hosted in wikikube Kubernetes behind the miscweb
ingress. This change updates to URLs to make sure requests are proxied
to the new endpoint in wikikube.

My knowledge of the wcqs request routing chain is limited, so this change
needs proper testing and eventually a fast rollback if the routing breaks.

Also rename the webserver-misc-eqiad mapping in trafficserver to
webserver-misc-sites.discovery.wmnet to make it less confusing, both cnames
point to miscweb1003.eqiad.wmnet.

Bug: T381909
Bug: T350793
Change-Id: I8f21160fa0094b547910e890aa7e453c74aad10a