dse-k8s: Use partman recipes for containerd with local storage support
In the dse-k8s cluster we have some nodes that have additional local
storage configured. We have a custom partman recipe for these, but it
needs adapting to make it compatible with the partition layout that is
desired with containerd.

This commit adds a new recipe and configures all of the dse-k8s-workers
to use containerd compatible recipes.

Bug: T377875
Change-Id: I4291498dd1d95c14208fd5030da90aace33acefb