dse-k8s: Stop installing the amd rocm packages to dse-k8s-worker1001
We wish to upgrade dse-k8s-worker1001 from bullseye to bookworm, but we
have not yet ported the amd rocm packages to bookworm.

This means that we won't be able to install these packages, which will
affect the monitoring of the GPUs in this host.

As we are not yet using the GPUs, the pragmatic decision is to remove
the requirement to install the amd rocm packages when reinatalling the
host. We can come back to the requirement later.

Note that this only affects the older generation of GPUs, not the newer
cards that are installed to dse-k8s-worker1009.

Hosts: dse-k8s-worker1001.eqiad.wmnet
Bug: T377875
Change-Id: Id6c58b3369c284b67d866287dfd9141fa50ca2d2