update dsek8s cluster to use containerd
we are in the process of migrating the dse-k8s cluster to use containerd instead
of docker. This patch sets up the
nrpe_check_disk_options to also exclude the containerd partition and the
puppet roles for control planes and workers to use container_runtime.
The hosts already on bookkworm are still set to docker and this shall be changed
as we reimage each host.

Bug: T377875
Change-Id: I7a5c553c36516810be067ac1fb3f711fdaa28b35