kernel-messages: add category=keyword_error
Some kernel error messages are logged with "priority=warning", but are
actually quite serious (see T386083). This commit adds a new dedicated
category for messages containing the word "error" that we can alert on,
without alerting for any message with "priority=warning".

Please note the keyword is applied to all priorities, even lower than
warning, just in case some errors are being logged with priority=notice
or lower.

I also added to the ignore list one message containing the word "error"
that is not an actual error. I found that message with this quick test:

    cumin cloudvirt* "journalctl --quiet -k -g error"

Bug: T386083
Change-Id: Ib5948d09fbea6342a77ab08623119e3b775e182c