puppetserver: fix puppet dir dependency issue in cloudvps masters
If the special parameter "ssldir_on_srv" is used, which is the case
for puppetmasters on cloud VPS, puppet attempts to create the directory
/srv/puppet/server.

But it does not ensure that /srv/puppet/ exists.

This leads to puppet dependency errors when puppet runs for the first
time on a new puppetmaster.

Error: Cannot create /srv/puppet/server; parent directory /srv/puppet
does not exist.

To fix this, replace ensure_resource with the wmf in-house function
mkdir_p, which can recursively create directory trees and is already
used in other places here in the same file.

This error showed up as part of T382960 when trying to create a new
local puppetmaster from scratch.

It should fix creating any new local puppetmasters across all cloud
VPS projects.

Bug: T382960
Change-Id: Ifc7f2353205c565a0ca2eb3232a687061ed500c3