deployment_server: Read mwscript-k8s MW image from values, not kube API
This is a little worse, because in edge cases when mw-web helmfile diffs
exist, it'll use the image that helmfile *would* use, not the image
actually running in production. But for now it's necessary for T378429,
because the existing solution requires access to the mw-web Kubernetes
configs (i.e. membership in the deployment group). This is a quick fix
that does the right thing almost always, and we might circle back and
cover the edge cases with something scap-integrated.

Bug: T378429
Change-Id: Idca1c364fb0f94f60eebf58b98b904b3772089db