gNMIc: Increase prometheus worker threads and cache time
I made these changes manually on netflow1002 yesterday and the
result was to fix the missing stats we were seeing in eqiad.

It may also be worth increasing the CPU count on the busier VMs
but these changes alone seem to have corrected the problem we
seen previously.

Removing the "writer_threads" paramater as that does nothing here,
it only relates to a prometheus remote-write API, not scrape-based
like we are using.  I added in error previously.

Bug: T386807
Change-Id: Ie5ed2fc42b190d52c40b807f4c1932ea0093ec8e