pontoon: add rolegroups functionality
A rolegroup represents a collection of roles and their settings (hiera
files) that can be shipped together. The intent is to make it easier to
get started, especially in new stacks.

All 'roles' listed will be added to the stack if not already present,
and their 'settings' list will be also added to the stack.

Additionally, a mechanism to pick default hostnames for VMs in added to
specmap.

Change-Id: I6046d6487e18bd21b2387151ad5b30009c4040c2