Rename text interface state values returned by GNMI to ints
The OpenConfig model for '/interfaces/interface/state/oper-status'
returns one of several text values to reference the current
operational state of an interface, corresponding to the states
defined in IF-MIB::ifOperStatus:

https://github.com/openconfig/public/blob/b0d7b808f02c99a5307368604a638d52b98dd593/release/models/interfaces/openconfig-interfaces.yang#L651

Similar to the issue we had with BGP status this causes a problem
when we try to export to prometheus as it only supports numeric
values.  This patch adds another 'event-strings' output processor
in gnmic to rename from the text values to those defined in the
model.

Bug: T372457
Change-Id: I9394ad8382cda7840591cf1200d2d5ab7fa442e3