gerrit: remove explicit UseG1GC flag
When Gerrit had memory trouble we ran Java 8 at the time and went to
enable G1 garbage collection (-XX:+UseG1GC). It eventually turned out to
be a memory leak (J232, T263008). Now that Gerrit is running Java 17,
the default garbage collector is G1 and we can remove the explicit
option, it is enabled by default:

  gerrit1003$ java -XX:+PrintFlagsFinal -version |& grep UseG1GC
      bool UseG1GC      = true             {product} {ergonomic}

Bug: T387223
Change-Id: I06044ea1e679a5d17f255f3ff8524634c11ccf86