Update CentralAuth multi-DC rules for SUL3, attempt 2
* Add Special:CentralLogin to exceptions. Like
  Special:CentralAutoLogin, it's used for redirect chains, often
  immediately after account creation/autocreation, so replication
  lag could be a problem.
* Remove loginwiki from exceptions. Special:CentralLogin and
  Special:CentralAutoLogin have it fully covered, it's not used
  in other ways during authentication.
* Do NOT add auth.wikimedia.org. Other than central login/autologin,
  normal POST rules should suffice there.

For SUL3, this leaves the GET request for returning from the shared
domain to the local domain, which is the return URL used by
AuthManager for remote providers, canonically Special:UserLogin/return.
Unfortunately due to localization its actual URL is unpredictable.
If this turns out to be a problem, we can artificially make it more
unique; for now, just hope it will be fine (not having
Special:CentralLogin in the exemption list until now didn't seem to
cause any problems either).

I02ed064473b50f0 documents the URL formatting requirements on the
CentralAuth side.

Test plan: find an IP that gets routed to the passive DC, run
  curl -vo/dev/null \
  'https://auth.wikimedia.org/enwiki/wiki/Special:CentralLogin' \
  |& grep x-cache
check wich cpXXXX host it returns, deploy the patch to that host,
verify (via the 'server' header) that now it's routed to the
active DC. Maybe repeat with Special:CentralAutoLogin, and with
https://en.wikipedia.org/wiki/Special:CentralLogin

This reverts commit 0459ef3a4e4e24b5877fc43e692f32b487e3abfe
and restores commit 014a838899e396fb32a262eb2913d92b533a3e7d.

Bug: T363695
Change-Id: I106fac80379552c795bc04c3ebdc8f74b785b09f