package_builder: use suite name n instead of archive name a in backports hook
The current pbuilder D02backports hook pins backports base on the
archive a=bookworm-backports (or bullseye, buster...).
However there is no archive with the name bookworm-backports, there is only a
archive named stable-backports. This leads to problems with installing
packages from backports during package builds (like golang 1.23 from backports)

There is a suite named bookworm-backports. So this change fixes the D02backports
hook to use the suite name instead of the achrive name.

I'm not fully sure if this is the proper way to fix this issue. The backports
hook could also use the proper archive name (stable-backports)?

Change-Id: I66d89e113a2dfef93d2bf12be9ef7bef77ee8831