mwscript: Do not run mesh checks in loops
Why:
In I67724eedd21d8456b762c8c94b304e6bef90584b we introduce a mechanism to
check that the service mesh is up before proceeding with the execution
of a maintenance script. This check is only necessary on the first
iteration.

What:
Flip the environment variable MESH_CHECK_SKIP to 1 once done with the
first call of the loop so the subsequent calls do not healthcheck the
service mesh.

Hosts: mwmaint2002.codfw.wmnet
Bug: T387208
Change-Id: I88aba0a8bb87ff58d917339d0477b6aec03f319a