trove.conf: change default volume to /dev/sdb
This is frustrating; different versions of ubuntu seem to default
to different default dev schemes, which means we're ping-ponging
a bit between names. This change will work for new trove instances
but may break existing/old instances.

At the very least, it seems good to have server and guest agent
agree on this. I suspect that the real fix is for trove to
get the dev name from cinder rather than coding it in a config.

Bug: T381959
Change-Id: Idc5b358531ca006db9805bade4884e95244a915e