mediawiki: add rewrite for rt.wikimedia.org to wikitech page
We recently decom'ed RT (Request Tracker), the old ticket system.

There will still be links to it in Phabricator and Gerrit.

We did not want to completely break the URLs but also not have
complicated rules mapping to every individual ticket since
most have been imported into Phabricator (Phorge) but with a variable
offset in the ticket numbers so there is no simple mapping it seems.

Instead we want to rewrite to the existing Wikitech page about RT
and then we can edit that to explain what happened and how to find
imported tickets.

Bug: T385777
Change-Id: I3e92543bbbc3a6483891232cff2ec33238551d97