community_civicrm: dovecot module for serving up local mail
Adds a dovecot server for use on the community_civicrm hosts. This
server is used to collect and serve VERP and bounce messages for civicrm
so they can be properly processed. It will only be available on the
localhost and over IMAP.

Will require a private repo config similar to:
    https://gerrit.wikimedia.org/r/c/labs/private/+/1124204

Bug: T383715
Change-Id: I66f3902d5a9df6234b8a6b13501fa78fe5c49b1a