deployment server: Don't pass -Dfull_image_build:True to scap stage-train
Rely on the existing logic in the image building code to
determine whether or not a full image build should occur.

This (hopefully) will result in better behavior in the case where
someone needs to re-run the presync command.  Passing
`-Dfull_image_build:True` during a re-run disallows for the
possibility of a quicker operation.

Bug: T387823
Change-Id: I8701825586294e8dfe8a28b7d847df80c3fecab0