dbbackups: Prepare backup1013 to take over eqiad backups of es* dbs
backup1013 is going to substitute backup1002 as the location where
external store backups are taken for eqiad. We will setup
the puppet and other code and define the new necessary grants, but
not enable any backup jobs yet, as not only they are new hosts,
but also will be an OS upgrade for this specific role, so there
could be some issues to tune afterwards.

Bug: T387892
Change-Id: I774a85293d8880059a0adb49122a3a970397ec98