dbbackups: Prepare backup2013 to take over codfw backups of es* dbs
backup2013 is going to substitute backup2002 as the location where
external store backups are taken for codfw. We will setup
the puppet and other code and define the new necessary grants, but
not enable any backup jobs yet, as not only they are new hosts,
but also will be an OS upgrade for this specific role, so there
could be some issues to tune afterwards.

Bug: T387892
Change-Id: Ic2e2f2360b98221bfd638fe6a351faa5ee8e1c0b