dbbackups: Migrate es backups from backup[12]02 to backup[12]13
Disable database backups generated at backup1002 and backup2002
and enable those jobs, with the same config, at backup1013 and
backup2013.

This is needed in order to generate External Store backups on the
newly setup hosts, and previous to decommission of backup1002 and
backup2002. It will also upgrade this stack to bookworm as a side
effect.

Bug: T387892
Change-Id: Idb0264fc8bf493dd35b2f91d33c56c318d47d1a7