Add component/lshw on bullseye
The cache hosts currently installed a version of lshw from bullseye-backports
(originally done in T380295).

backports is only available for Debian stable, so bullseye-backports is
no longer being updated and will be archived at some point (which would
break Puppet runs and reimages).

As such, create a separate repository component/lshw to import the current
version from bullseye-backports. lshw is otherwise installed fleet-wide
and it's unclear what the backwards compat looks like, so to avoid subtle
mistakes it seems better to not copy the updated lshw to "main", but to
confine it to a repository component only used on the cache hosts.

Bug: T380295
Bug: T383557
Change-Id: I468c7d4c5b6be2c30c510b4c1dce28f46f025949