Install lshw backport from component/lshw
The cache hosts currently installed a version of lshw from bullseye-backports
(originally done in T380295).

backports is only available for Debian stable, so bullseye-backports is
no longer being updated and will be archived at some point (which would
break Puppet runs and reimages).

As such, install them from a separate repository component/lshw.

lshw is otherwise installed fleet-wide and it's unclear what the backwards
compat looks like, so to avoid subtle mistakes it seems better to not copy
the updated lshw to "main", but to confine it to a repository component
only used on the cache hosts.

Bug: T380295
Bug: T383557
Hosts: cp1114.eqiad.wmnet
Change-Id: I08f6a845fc5a2ecbb60cc5d013c55b47d16f2766