Remove sudo privileges for journalctl from airflow instance admins
This sudoers entry for allowing airflow instance admins to read their
instance's log files was unnecessarily relaxed in its use of the
wildcard shell glob.

We do not need the functionality, since logs are also written to the
file system with the correct file privileges. Also these VMs will be
deprecated and decommissioned soon, so it is not worth the effort to
develop a wrapper script for making this easier and safer for users.

Bug: T387719
Change-Id: I56c9c47067a3f38125bb2263c4bed4e6bd75b705