dbbackups: Prepare backup1002, backup2002 for decommissioning
Set backup1002.eqiad.wmnet and backup2002.codfw.wmnet as insetup,
so content there is still available, but no further backup work
happens there.

We will keep the host untouched for a few weeks to confirm the
bookwork backup workflow of the new hosts that substituted them
(backup1013 and backup2013) work as expected and then send them
for decommissioning.

This requires deletion of grants deployment to mariadb on several
hosts, in addition to puppet.

Bug: T387892
Change-Id: Ib63bbf711637268ee4145e02f0eb3e29c444dd09