Profile::Mediawiki_deployment: add 'deploy' field to release config
The 'deploy' field is an optional boolean (defaults true in scap), that
controls whether scap manages deployents of the release. If false, it
manages only the /etc/helmfile-values/mediawiki/release values file.

Also adds a basic file header comment, including a callout to the
DeploymentsConfig docstring for details on hash fields.

Bug: T387917
Change-Id: If1aa9ee358a631a40c3ef68bf816f5198dac43f8