keepalived: Install keepalived from the "main" component
backports is only supported for the life time of Debian stable/oldstable,
but not LTS and it will be archived at some point.

keepalived on bullseye is only used on the cloudgw hosts, so prior
to merging this patch I would import keepalived 1:2.2.7-1~bpo11+1
to the "main" component of bullseye-wikimedia.

While we're at it, also remove obsolete root:root annotations for
the file resource, this is the default since
7acc980cf4f5e2166a531dccb4f7c30f1cffe589

Bug: T383557
Change-Id: I4c4a5a3c7f34baf34867d69f8b8eb53634a65724