trafficserver: revert cookie-enrolled traffic to main
With main now on 8.1, stop diverting cookie-enrolled traffic to next.

I'd originally imagined doing this in a couple of steps, rather than
going to load_fraction = 0 all at once, but I'm not sure that's worth
the hassle given how much traffic is involved.

Instead, I'll splay this out over ~ 30m or so when applying the patch to
A:cp-text hosts via cumin.

Note: To be merged *after* Ib56dd24c15899a89477dd46f44ee6f34186152f7 is
merged and applied.

Bug: T383845
Change-Id: I4efb7610cd478a345962d258cbfacf418741c0d2