haproxy: certificate check script
A very simple bash script to use as prerequisite for the haproxy
systemd unit to check if certificates are expired. The certificate
list to validate is the same used for haproxy tls configuration, but
wrote to a different file that is passed as parameter to the script
from the systemd unit. This makes the script more generic and useful
to check other certificates too.

Hosts: cp1111.eqiad.wmnet,cp4037.ulsfo.wmnet
Bug: T388147
Change-Id: I05a30e9967b9abe1d09f28f5cbb8c17e49df7783