Update the version of refinery used for refine_sanitize jobs
The use of the previous versions was causing the refine_sanitize jobs
to use a deprecated SMTP mail exchange server, rather than the new
default settings of localhost:25.

We update the refinery jar for both the test and production instances of
refine_sanitize jobs in this patch.

Hosts: an-launcher1002.eqiad.wmnet
Bug: T388417
Change-Id: I0b4ce812aa3ce1fef9f59868d68ea15b1e5387e5