deployment_server: Support PHP version selection in mwscript-k8s
With this change, mwscript-k8s can now target a specific PHP version to
use, by way of selecting the appropriate release values file based on
the --php_version flag (initially defaults to 7.4, so no change).

Once the default is switched to 8.1, this can be used as an escape hatch
by users who in encounter compatibility issues with their scripts.

Bug: T387917
Change-Id: I2d95e3c0d69559e3a37063a82689d640f692b9c9