lists/stewards: allow switching some lists to actual "wet"-run
So far the system for automating list member syncs for steward-related
lists had all lists set to just a dry-run where we see what _would_
happen without actually making changes.

Now the system is ready to switch the first list, stewards-l, to
actually run the sync while other lists should stay with the dry runs.

https://en.wikipedia.org/wiki/Dry_run_(testing) says the term dry
run comes from firefighting and the opposite was actually called
a "wet run", so we are going to use that term here.

Switch stewards-l to actually running the sync command.

Bug: T388354
Change-Id: Ib6046e68971995f048c2f0fa7945eee36b4dbb8b