wikidough: add healthcheck override for doh1001 and doh2002
These two hosts were upgraded to dnsdist 1.9.8, which uses nghttp2 for
the DoH frontend and only supports HTTP/2. But check_http doesn't
support that so the current healthcheck fails.

A proper way to fix this is to use check_curl but that is also failing
(T388680). For now, we accept the 400 response to confirm that at least
dnsdist is running.

Hosts: doh1001.wikimedia.org, doh1002.wikimedia.org, doh2002.wikimedia.org
Change-Id: I530815c427bd9cc3a61edf0ab1a9d36651eeab6e