osm_master: Tighten replication slot name on Bookworm
When debugging why replication between maps-test2001 and maps-test2002
failed I noticed the following in the logs on maps-test2001:

  Notice: /Stage[main]/Postgresql::Master/Exec[create slot postgresql::
  master_wal_maps-test2002_codfw_wmnet]/returns: ERROR:  replication slot
  name "wal_maps-test2002_codfw_wmnet" contains invalid character
  Notice: /Stage[main]/Postgresql::Master/Exec[create slot postgresql::
  master_wal_maps-test2002_codfw_wmnet]/returns: HINT:  Replication slot
  names may only contain lower case letters, numbers,
  and the underscore character.

Poking at the logs the slot name passed was wal_maps-test2002_codfw_wmnet.

So it seems that between Postgres 11 and 15 the syntax is now enforced
more strictly (or maybe it was never really enforced in 11, who knows).

In order to not mess with the existing maps cluster, add an OS conditional
to only apply the new syntax for the Bookworm hosts.

Hosts: maps2009.codfw.wmnet,maps1009.eqiad.wmnet,maps-test2001.codfw.wmnet
Bug: T381565
Change-Id: Ib945d9a2963015111091a48f8442efaa126c13e3