service: move kartotherian-k8s-ssl fully on k8s
The kartotherian-k8s-ssl VIP is currently configured to be able
to serve traffic from both bare metal (maps*) and Wikikube
workers. We have completely depooled bare metal hosts a week ago,
not traffic is flowing to them anymore, so the VIP can now be
moved to a proper k8s config (like the other ones).

Bug: T386926
Change-Id: I42da3539ae5fac03fa829b038a1e7a0085fd102a