opensearch: Symlink sudachi dictionary into per-instance config
The sudachi language analyzer plugin used in the search platform
clusters installs a dictionary, but due to the way the plugin is
designed that dictionary needs to be available in the config
directory rather than the plugins directory.

If the dictionary is available then symlink it into each instance
config directory.

Hosts: cloudelastic1012.eqiad.wmnet
Change-Id: I7ee7e9fa7706610e3658391da929f458f7ee52d8