mailman: list sync, add option to mail changes to an admin
Create option and parameters to mail the output
of list member sync commands to an external admin.

For auditing purposes. So that in this use case Martin
Urbanec gets an email when stewards-l members are being
synced.

The output will either be "Nothing to do" or additions/deletions
of members on the mailing list.

As a response to T351202#10616848.

Bug: T351202
Change-Id: Ie41ecb04b20befbf9bdd9d3b9d4a99e0c8a42430