swift: remove ms-be2075 from rings
This host has been drained for a while due to h/w issues, now resolved
(cf. T382707). So it makes a useful test host for the VLAN moves
required in T354872.

The VLAN move will renumber the host, however, so it needs to be
entirely removed from the rings (rather than just drained) before
being moved, and then re-added with the new IP address.

Bug: T354872
Change-Id: I87b88dd89ed113d03d0d84da666062291579624e