spicerack: add Spicerack interactive shell
Why:
See the related task for all the context on the why.

What:
Implement a quick script that uses `python -i` to drop into a REPL
interactive shell with a spicerack instance pre-initialized and logging
both on screen and on file setup like in the cookbooks.
This will allow to run both live and dry-run lines of code that use
spicerack and discover/test them more in isolation.

Hosts: auto
Bug: T389329
Change-Id: Id792a3f23cd915dfa155a01ec31a0a444fdc4f17