haproxy: using tmpfs directory for private tls material
Allow user to specify if TLS private keys should be stored in tmpfs
directory.

This required patching both `sslcert` and `acme_chief` module to
support optional directory parameter where to store TLS material
instead of hardcoded path.

The hiera key `profile::cache::haproxy::use_tls_tmpfiles` is used to
control this behavior ensuring that when set to false the certificates
and keys keeps being stored in the non-volatile paths, allowing for
safer rollout of this feature.

An additional check is used to ensure that if the use_tls_tmpfiles
switch is true, certificates in hiera files must comply with the
correct path (volatile or non-volatile).

Currently this is brought only on one (depooled) host for test.

Hosts: cp1111.eqiad.wmnet,cp4047.ulsfo.wmnet
Bug: T384227
Change-Id: Icd1c1b95648e6990d28bfe3e985b7124c27ab6c8