Allow analytice-wmde-users limited journalctl access to their units
We need to allow members of the analytics-wmde-users group access to
journalctl in order to allow them to debug the current failure of some
graphite based analytics scripts.

There is an existing rsyslog fragment that is intended to write the
logs to disk, but it is not working at present.

Bug: T387514
Change-Id: I82d35f84cae65488d72725998d5219d814ae6d87