Spiderpig: Declare the spiderpig group
Why:
The puppet user resource issues

    /usr/sbin/useradd -c SpiderPig jobrunner/apiserver 
    -g 929 
    -d /var/lib/spiderpig 
    -u 929 
    -M spiderpig
    
as a command to create the user.
However, -g expects the group to already exist.

What:
Declare the group with the proper gid and set as system

Note:
Given deployment servers are still bullseye, we can't use
systemd-sysuser as it is still plagued by some libc related bugs. Put a
comment to move to systemd-sysuser once we are past bookworm

Hosts: auto
Bug: T383945
Change-Id: I086e444da0564cbe8ee1370ca25cbd748feb2294