spiderpig: Switch to listening on :: instead of 0.0.0.0
Why:
The default of 0.0.0.0, means only listening on IPv4. Since the
deployment hosts have IPv6 AAAA records for some time now and the
hostname is defined as the destination for the TLS terminator rule,
envoy tries out the IPv6 connection, which fails. Instead of falling
back to IPv4 it returns an error.

What:
Switch SpiderPig APIserver to :: instead of 0.0.0.0 as this is faster,
easier and long term future proof, at least compared to fighting envoy
ERB templating to convince envoy to fallback to IPv4

Hosts: auto
Bug: T383945
Change-Id: I74847c044eba450266dded6a7bac1da47d21b5f2