varnish: Fix X-Wikimedia-Debug cookie handling
See https://wikitech.wikimedia.org/wiki/WikimediaDebug

These cookies typically take the form of
`backend=mwdebug1001;log;expire=1742415486` (which gets URL encoded
in the Cookie header, unlike the X-Wikimedia-Debug header which is not
encoded). The VCL logic decodes the cookie value to convert it to the
header, but accidentally stopped at the first ; or = sign.

Bug: T350094
Change-Id: Ic9687d9fe984e135c8a45ea85af9199be036ed9c