Profile::Mediawiki_deployment: add mw_kind fields
This adds mw_kind (i.e., mediawiki image kind) at the namespace- and
per-release-override level.

Since debug will also be deprecated in favor of mw_kind: debug-image and
stage: testservers, add a TODO to remove the former and expand the enum
value range for the latter to cover all stages.

Bug: T389499
Change-Id: I2c2bd756a8fde933cbd06eee79f846bdf9f98cfb