dynamicproxy::api: Install python3-flask-sqlalchemy from "main" component
backports is only supported for the life time of Debian stable/oldstable,
but not LTS and it will be archived soon.

python3-flask-sqlalchemy on bullseye isn't used anywhere else, so prior
to merging this patch I imported python3-flask-sqlalchemy 2.1-4 to the
"main" component of bullseye-wikimedia.

While we're at it, also remove obsolete root:root annotations for
the file resources, root:root is the default since
7acc980cf4f5e2166a531dccb4f7c30f1cffe589

Bug: T383557
Change-Id: I5668d0fd07c9be26e0ee3aac3a557ab8861cca5a