webperf: Move `php_admin_flag engine on` from subdir to docroot
Today PHP is enabled on docroot/excimer and docroot/xhgui.
We got here because it started as a fully static site with /excimer
and /xhgui proxied to webservers on other hosts where PHP was enabled.

When we folded these into one, we migrated it exactly with PHP
enabled only in these two subdirs. The top-level site remains served
as static HTML from performance/docroot.git:public_html where each
commit performs a build.

I'd like to explore simplifying this:

1. Move the tiny PHP transformation from a pre-commit build step to
   happen at runtime in the public_html directory.
2. Put /xhgui and /excimer in their correct place in public_html.
3. Remove /xhgui Alias from here, as it will be redundant by then.
   This will ease local testing and URL navigation by working
   out-of-the-box.

Change-Id: I4321211851a6aab8449aa2edadf876f79e76b118