postgresl/osm_master: Make postgresql listen on ipv4 on Bookworm
The legacy maps stack on Buster only used ipv4 addresses on the servers.

This changed with the new maps bookworm stack, but with Postgres 15
the client side tooling prefers ipv6 connections, which leads to
connection issues since our ACLs are host-based and solely based on
ipv4 IPs.

I'll file a task to eventually migrate them to FQDN-based rules, but
that's a bigger task and for now we'll have postgresql only listen
on ipv4.

Hosts: C:postgresql::master
Bug: T381565
Change-Id: I3a2601a3aa7aa99ace9732d35af644935ef5538c