Configure a scap deployment of mediwiki-dumps-legacy
We wish for scap to be able to deploy mediawiki to the
mediawiki-dumps-legacy namespace on the dse-k8s-eqiad cluster.

Although no pods will be created, a suspended job template will be
deployed, which will subsequently be used by Airflow jobs.

The mediawiki-cli image variant will be selected and added to the
helmfile values file supporting this deployment.

Hosts: deploy1003.eqiad.wmnet
Bug: T389786
Bug: T389499
Co-Authored-by: Balthazar Rouberol <brouberol@wikimedia.org>
Change-Id: I8161831a8d1a74ceb96a1a54170940ba652fde6a