Allow releng to resume train related systemd timers
On Tuesday, the deployment server runs two scap commands to prepare the
MediaWiki train: train-presync and train-clean. When those tasks fail,
we investigate the failure and then have to manually rerun the tasks.

We can copy paste the command which is prone to error. This change adds
sudo rules to allow the train conductors (release engineering) to start
the timers manually and inspect the logs.

Bug: T387823
Change-Id: I09f89e59276f60018166b3f7cf16b6e918c29f74