Create insetup role for SRE Collab with nftables and rename existing one
Currently all our insetup roles default to Ferm, but with an increasing
amount of roles defaulting to nftables this is becoming problematic:
It means these are installed with Ferm initially and then Puppet
applies the switch to nftables. But this switch requires a reboot (to
unload the iptables kernel modules), which might be missed, causing
issues later on.

As such make the choice of the firewall stack part of the role, so
that it can be explicitly selected when creating the racking task
for DC ops:
- rename the existing role to _ferm
- create a new role _nftables which defaults to nftables

The insetup hosts in site.pp which correlate to a role which is already
switched to nftables are fixed alongside.

Bug: T389825
Change-Id: Idca5821b57d67cee0c2f145491d695ab571076d6