elastic/cirrussearch: begin production migration
We are migrating the production Elasticsearch cluster
to OpenSearch, and renaming the hosts from "elastic"
to "cirrussearch". We'll begin in CODFW and go row-by-row,
so I've arranged the current elastic regexes in
site.pp by row.

I've also added the new hostnames with their new roles
as comments.

Hosts: re:elastic2.*
Bug: T388610
Change-Id: Ic07d547a23fe63486a7aabd5eae957b09b6db79f