data.yaml: Fix journalctl access to spiderpig services
This is a followup to Ic1071c59ca041eae13149db5ac2c1d15ddd55764

The `*` at the end of the sudo entry makes it so the user is required
to supply parameters, which is confusing.  Also, accepting arbitrary
additional parameters allows undesirable access to logs of other
services.

Hosts: deploy1003.eqiad.wmnet, deploy2002.codfw.wmnet
Bug: T383945
Change-Id: Icefdf397419d503d1106b322b695eed217a0cf05