Add JsonConfig's globaljsonlinks tables to catalog
JsonConfig recently added globaljsonlinks, globaljsonlinks_wiki,
and globaljsonlinks_target tables to track usage of Data: pages across
wikis, living in production on x1.commonswiki.

There is also a test instance in the db for testcommonswiki, which may
at some point in the future be removed.

Added to the machine-readable table catalogue, along with the link to
the schema defining them.

Bug: T363581
Change-Id: Ide73303fd61e0bc524773a359fa6e733ea0ea8f6