benthos: update the webrequest_live instance
The DE team is ready to move webrequest away from varnishkafka to
haproxy, the two streams have been running in parallel for a while
and DE is happy about the consistency checks.

The two streams are identical except for the new one that brings
the `server_pid` field, but DE wants to keep it so I think
there is nothing to do in this change.

This change:
* Moves the source topics to webrequest_frontend_{text,upload}
* Changes the Kafka consumer group as well, to start clean/fresh.
* Removes the logic for the handling of `dt: "-"` since it was
  a peculiar case of varnishkafka, it doesn't happen for haproxy.

The alternative road is to create another webrequest sampled instance,
test it in parallel with the current one, and make a final switch.
It may be heavy on the central log hosts, and cumbersome to manage,
but safer. The worst case scenario for this one is that we rollback
and we loose some data (or maybe we have some minutes of inconsistency),
but given the nature of the stream it should be fine.

Bug: T390029
Change-Id: I7f61391eeb90afe451f7955846b21f8bec43de65