deployment_server: Default to PHP 8.1 in mwscript-k8s
This switches the default PHP version to 8.1.

If the user provides neither --mediawiki_image nor --php_version, log a
message noting that the job will use 8.1 and --php_version can be used
to select 7.4.

If the user provides --php_version 7.4, log a message noting that this
functionality will be removed, urging the user to fix the underlying
compatibility problem.

Bug: T387917
Change-Id: Ide16e37aea447285ec6ca261cd688d981d68087f