Create insetup role for ML servers with nftables and rename existing one
Currently all our insetup roles default to Ferm, but with an increasing
amount of roles defaulting to nftables this is becoming problematic:
It means these are installed with Ferm initially and then Puppet
applies the switch to nftables. But this switch requires a reboot (to
unload the iptables kernel modules), which might be missed, causing
issues later on.

As such make the choice of the firewall stack part of the role, so
that it can be explicitly selected when creating the racking task
for DC ops:
- rename the existing role to _ferm
- create a new role _nftables which defaults to nftables

There are currently no ML roles defaulting to nftables, so this is mostly
prep work and bringing it in sync with the rest.

Bug: T389825
Change-Id: Ie0c5d6b3f677cf5526fae561a11e45e4c1fc0622