cirrussearch: use EFI for soon-to-be-reimaged Elastic hosts
As we've had issues reimaging our elastic hosts using BIOS
and HTTP together in the past, we're going to switch to EFI
so we can use the faster and more reliable HTTP with the newer
EFI boot option. We've used this on a few hosts already with no issues.

Hosts: O:installserver
Bug: T388610
Change-Id: I81b1ee1cae2569b0f8324ad57544326358d4dcee