trafficserver: gateway-check ignore list, roll pcs/mobileapps to more wikis
We would like to further route mobileapps-direct traffic (ie without
using restbase) to all but a subset of wikis that are either currently
subject to fundraising campaigns and enwiki due to its overall size.

In order to do this in changeprop, we use a negative lookahead regex,
but lua lacks a feature complete set of regexes, so in order to
accomplish similar in our trafficserver plugin we define an ignore
list where a wiki and a list of URLs can be defined as suitable for
ignoring, and simply specify the wikis to avoid. This feature may be
useful for other migrations, or it might be something we can simply
remove once the PCS/mobileapps migration is complete.

Bug: T388140
Change-Id: I94535d3ea7eab1ce414923b980f0b47ebfbaf4dc