elastic: enable performance governor for selected hosts
While migrating servers in T388610, we noticed some
increased load and latency on the cluster. This
reminded me of T386860, where we discussed enabling
the performance governor on our R450 hosts.

I've reviewed the power usage on the DC using
the dashboard at https://w.wiki/Dbdu , and this patch
enables the performance governor for CODFW hosts
that are NOT in the top 3 power-consuming racks.

Hosts: elastic2083.codfw.wmnet,elastic2089.codfw.wmnet
Bug: T388610
Bug: T386860
Change-Id: I103c74f680bdb7f5bdc45b14ff7c4779bd1f926b