Create insetup role for SRE o11y with nftables and rename existing one
Currently all our insetup roles default to Ferm, but with an increasing
amount of roles defaulting to nftables this is becoming problematic:
It means these are installed with Ferm initiaally and then Puppet
applies the switch to nftables. But this switch requires a reboot (to
unload the iptables kernel modules), which might be missed, causing
issues later on.

As such to make the choice of the firewall stack part of the role, so
that it can be explicitly selected when creating the racking task
for DC ops:
- rename the exisiting role to _ferm
- create a new role _nftables with defaults to nftables
- rename existing entries in site.pp

There are currently no o11y insetup hosts, so nothing needs to be
updated in site.pp

Bug: T389825
Change-Id: Ibd94394637ee4ce338f5af444547f37d2dd5d0cc