wdqs: fix monitoring user-agents
in Ic77828c3669c151525898ccca2bbee18de046513 the blackbox check for all
wdqs endpoints changed its UA to prometheus-sparql-check.
Adjust the list of monitoring UAs so that monitoring requests are
properly flagged as such.
Flagging monitoring query is important to determine whether an host is
online or not so that it does not interfere with wikidata maxlag if it's
explicitely depooled to catchup on lag.

Change-Id: I274bd8795e922aca57e596af3dbcc374c40af1be