Add a cleanup timer for old dumps webrequest logs
We copy the log files from the clouddumps100[1-2] servers to stat1011
for analysis to gather metrics around wikidata entity dump usage.

However, we have not had a housekeeping method in place to remove the
old files, so they are retained on disk for longer than intended.

This commit adds a systemd timer that removes log files older than 90
days.

Hosts: stat1011.eqiad.wmnet
Bug: T390123
Change-Id: I9a1b1a035f6d181b429376a15b82531d1fc37a71