Create insetup role for Data Platform with nftables and merge DE/Search roles
Currently all our insetup roles default to Ferm, but with an increasing
amount of roles defaulting to nftables this is becoming problematic:
It means these are installed with Ferm initiaally and then Puppet
applies the switch to nftables. But this switch requires a reboot (to
unload the iptables kernel modules), which might be missed, causing
issues later on.

As such to make the choice of the firewall stack part of the role, so
that it can be explicitly selected when creating the racking task
for DC ops:
- Merge the existing insetup roles for Data Engineering and Search into
  insetup::data_platform_ferm
- create a new role insetup::data_platform__nftables which defaults to
  nftables
- rename existing entries in site.pp depending on role config
- remove matomo1002 entry in site.pp (leftover)

Bug: T389825
Change-Id: I622694392a4b5656633a3d320920c76da978d6e8