Upgrade PHP on the misc dumps worker - snapshot1016
We recently upgraded PHP on all of the dumps generation hosts, with the
exception of snapshot1016. The reason for this is that it was generating
an important wikidata entity dump at the time.

Now we wish to bring this last host in line with the others by upgrading
PHP on it, too.

Hosts: snapshot1016.eqiad.wmnet
Bug: T382484
Bug: T319432
Change-Id: Ic6e5809ee25db21fb67b4a7bf136023931fb6579