Update opensearch-madvise call for version 0.2
Version 0.2 of the opensearch-madvise program adds a new required
argument specifying the data path that it should operate within.
Update the timer to invoke it with the configured data path for
each cluster.

The package itself is marked `ensure => present`, as such when
this is deployed we will also need to issue the appropriate
cumin commands to install the updated package.

Bug: T390592
Hosts: cloudelastic1009.eqiad.wmnet
Change-Id: I78b27dd4e517566c511f91bc6ab1c96eea35e9b4