Set envoy keepalive's for search to match nginx
On the nginx side of these connections we have set the
keepalive_timeout to 60. But on the envoy side this is
unconfigured. I suspect this has led to a low level (~5/minute)
of `upstream connect error or disconnect/reset before headers.
reset reason: connection failure` errors.

Set an explicit keepalive timeout in envoy of 50s, a bit less
than what nginx is configured for.

Bug: T390612
Hosts: deploy1003.eqiad.wmnet
Change-Id: I5d723fa6e8b6157b8efc08d7eef61b540c0a0c14